+++ examples/js/sort/demo.js
@@ -11,6 +11,7 @@ import SortHookTable from './sort-hook-table';
 import DisableSortIndicatorTable from './disable-sort-indicator-table';
 import CustomCaretSortTable from './custom-caret-sort-table';
 import ExternalMultiSort from './manage-multi-sort-external-table';
+import DefaultASCSortTable from './default-asc-sort-table';
 
 class Demo extends React.Component {
   render() {
@@ -112,6 +113,15 @@ class Demo extends React.Component {
             <div className='panel-body'>
               <h5>Source in /examples/js/sort/custom-caret-sort-table.js</h5>
               <CustomCaretSortTable />
+            </div>
+          </div>
+        </div>
+        <div className='col-md-offset-1 col-md-8'>
+          <div className='panel panel-default'>
+            <div className='panel-heading'>Default ASC as First Sort Order</div>
+            <div className='panel-body'>
+              <h5>Source in /examples/js/sort/default-asc-sort-table.js</h5>
+              <DefaultASCSortTable />
             </div>
           </div>
         </div>