+++ src/TableHeaderColumn.js
@@ -1,6 +1,7 @@
 /* eslint default-case: 0 */
 /* eslint guard-for-in: 0 */
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import classSet from 'classnames';
 import Const from './Const';
 import Util from './util';
@@ -25,7 +26,9 @@ class TableHeaderColumn extends Component {
 
   handleColumnClick = () => {
     if (this.props.isOnlyHead || !this.props.dataSort) return;
-    const order = this.props.sort === Const.SORT_DESC ? Const.SORT_ASC : Const.SORT_DESC;
+    let { sort: order } = this.props;
+    if (!order && this.props.defaultASC) order = Const.SORT_ASC;
+    else order = this.props.sort === Const.SORT_DESC ? Const.SORT_ASC : Const.SORT_DESC;
     this.props.onSort(order, this.props.dataField);
   }
 
@@ -270,7 +273,8 @@ TableHeaderColumn.propTypes = {
   tdAttr: PropTypes.object,
   tdStyle: PropTypes.object,
   thStyle: PropTypes.object,
-  keyValidator: PropTypes.bool
+  keyValidator: PropTypes.bool,
+  defaultASC: PropTypes.bool
 };
 
 TableHeaderColumn.defaultProps = {
@@ -305,7 +309,8 @@ TableHeaderColumn.defaultProps = {
   tdAttr: undefined,
   tdStyle: undefined,
   thStyle: undefined,
-  keyValidator: false
+  keyValidator: false,
+  defaultASC: false
 };
 
 export default TableHeaderColumn;