+++ src/filters/Select.js
@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import classSet from 'classnames';
 import Const from '../Const';
 
@@ -57,11 +58,13 @@ class SelectFilter extends Component {
 
   getOptions() {
     const optionTags = [];
-    const { options, placeholder, columnName, selectText } = this.props;
+    const { options, placeholder, columnName, selectText, withoutEmptyOption } = this.props;
     const selectTextValue = (selectText !== undefined) ? selectText : 'Select';
-    optionTags.push((
-      <option key='-1' value=''>{ placeholder || `${selectTextValue} ${columnName}...` }</option>
-    ));
+    if (!withoutEmptyOption) {
+      optionTags.push((
+        <option key='-1' value=''>{ placeholder || `${selectTextValue} ${columnName}...` }</option>
+      ));
+    }
     Object.keys(options).map(key => {
       optionTags.push(<option key={ key } value={ key }>{ options[key] + '' }</option>);
     });